redpanda: split out debug route registration
